some clue about writer

+++ util/vector.h
@@ -18,17 +18,17 @@ void vec_fini(vector_t *vec);
 size_t vec_size(vector_t *vec);
 size_t vec_num(vector_t *vec);
 
-/* 没有默认值,get之后的数据有可能是随机 */
+/* 对于没有添加的值,为NULL */
 const void *vec_get(vector_t *vec, size_t offset);
 /* 
  *
- * 因为vec需要连续,所以value可以为NULL,删除需要显式删除
+ * value可以为NULL,但是无法区分默认为NULL还是设置为NULL
  * 返回值是添加的个数
- * -1表示vec为NULL
  * 0表示没有添加成功
  *
  */
-int vec_set(vector_t *vec, size_t offset, const void *value);
+size_t vec_set(vector_t *vec, size_t offset, const void *value);
+/* 删除操作会移动底层数据,和单纯置NULL不同 */
 const void *vec_erase(vector_t *vec, size_t offset);
 
 const void **vec_data(vector_t *vec);
@@ -40,7 +40,7 @@ const void **vec_data(vector_t *vec);
  */
 
 /* 实现原因,只优先支持末尾操作 */
-int vec_push(vector_t *vec, const void *value);
+size_t vec_push(vector_t *vec, const void *value);
 const void *vec_pop(vector_t *vec);
 const void *vec_front(vector_t *vec);
 const void *vec_back(vector_t *vec);
@@ -53,7 +53,7 @@ const void *vec_back(vector_t *vec);
  *
  */
 #define vec_append(vec, args...) vec_append_private(vec, ##args, NULL)
-int vec_append_private(vector_t *vec, ...);
+size_t vec_append_private(vector_t *vec, ...);
 
 #ifdef __cplusplus
 }
